+++ back-end/src/test/java/admin/worker/application/WorkerServiceTest.java
@@ -1,0 +1,20 @@
+package admin.worker.application;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+import org.junit.jupiter.api.Test;
+
+class WorkerServiceTest {
+
+    @Test
+    void popJobByServerId() {
+    }
+
+    @Test
+    void changeJobStatus() {
+    }
+
+    @Test
+    void updateWorkerStatus() {
+    }
+}
