+++ back-end/src/test/java/admin/worker/ui/WorkerControllerTest.java
@@ -1,0 +1,20 @@
+package admin.worker.ui;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+import org.junit.jupiter.api.Test;
+
+class WorkerControllerTest {
+
+    @Test
+    void takeJob() {
+    }
+
+    @Test
+    void updateJobStatus() {
+    }
+
+    @Test
+    void updateWorkerStatus() {
+    }
+}
